-Added basic texure system     -added texture class     -added texture handling to batch renderer -Rewrote fragment shader -added stb_image library -finished batch renderer -changed vertex.m_textureID to be an int instead of a float -added full virtual getTextures() method to renderable abstract class

+++ GalaxySim2020/src/Graphics/Primatives/vertex.h
@@ -8,6 +8,6 @@ namespace Graphics
 		glm::vec3 m_pos;
 		glm::vec4 m_color;
 		glm::vec2 m_texCoords;
-		float m_textureID;
+		int m_textureID;
 	};
 }
